pathfind, collision and simple graphic

+++ NSW_api/ext/includes/NSW_api/EUtils.cpp
@@ -9,7 +9,7 @@ float EMath::clamp_value_float(float _v, float _min, float _max)
 
 }
 
-float EMath::clamp_value_int(int _v, int _min, int _max)
+int EMath::clamp_value_int(int _v, int _min, int _max)
 {
 	if (_v < _min) { return _min; }
 	if (_v > _max) { return _max; }
