+++ NSW_api/ext/includes/NSW_api/EUtils.h
@@ -3,5 +3,5 @@ class EMath
 {
 public:
 	static float clamp_value_float(float _v, float _min, float _max);
-	static float clamp_value_int(int _v, int _min, int _max);
+	static int clamp_value_int(int _v, int _min, int _max);
 };
